Add a comment on default-features=false requirement
This is documented, but having it here as well can be useful.

Having "optimal" as a default feature is a workaround for Cargo's
inability to select dependencies by feature. For that workaround
to be less annoying to use, the "pure-rust" and "optimal" targets
should be mutually exclusive. The ability for Cargo to do that
was asked in 2016 but, just like the ability to select dependencies
by feature, the issue is still open today.